+++ portfolio/package.json
@@ -1,7 +1,7 @@
 {
   "name": "portfolio",
   "version": "0.1.0",
-  "homepage": "https://nulula.github.io/react-portfolio",
+  "homepage": "https://nulula.github.io/",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.5",
@@ -11,6 +11,7 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.8.2",
     "react-scripts": "5.0.1",
+    "react-scroll-parallax": "^3.3.2",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
