+++ portfolio/src/App.css
@@ -2,6 +2,12 @@
   text-align: center;
 }
 
+.home {
+  max-width: 100%;
+  margin: 0;
+  padding: 0;
+}
+
 .App-logo {
   height: 40vmin;
   pointer-events: none;
