+++ portfolio/src/App.js
@@ -1,31 +1,19 @@
 import logo from './logo.svg';
 import './App.css';
 import React from 'react';
-import Contact from './components/Contact';
+import Contact from './components/pages/Contact';
 import Header from './components/Header';
-import Home from './components/Home';
-import Projects from './components/Projects';
+import Home from './components/pages/Home';
+import Projects from './components/pages/Projects';
 import { Routes,Route } from 'react-router-dom';
+
 
 function App() {
   return (
-
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
+  <div>
+    <Home/>
+    <Projects/>
+  </div>
   );
 }
