--- portfolio/src/components/Contact.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-function Contact() {
-
-};
-
-
-
-export default Contact;
