--- portfolio/src/components/Home.js
@@ -1,8 +1,0 @@
-import React from 'react';
-
-function Home() {
-
-};
-
-
-export default Home;
