--- portfolio/src/components/Projects.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import Project from './Project';
-
-function Projects() {
-
-};
-
-
-
-export default Projects;
-
