+++ portfolio/src/components/pages/Contact.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function Contact() {
+
+};
+
+
+
+export default Contact;
