+++ portfolio/src/components/pages/Projects.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import Project from '../Project';
+
+function Projects() {
+
+};
+
+
+
+export default Projects;
+
